Statement-1-Insurance-Decoder/src/requirements.txt
streamlit
langchain
langchain-groq
langchain-huggingface
langchain-community
sentence-transformers
chromadb
pymupdf
python-dotenv
fastapi
uvicorn
pydantic
python-multipart
